backend/requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
Click==7.0
colorama==0.4.3
cryptography==2.8
dbus-next==0.2.3
defer==1.0.4
defusedxml==0.6.0
distlib==0.3.2
distro==1.4.0
Django~=3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==3.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
multidict==5.1.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.4.0
pkgconfig==1.5.5
platformdirs==2.3.0
protobuf==3.6.1
psycopg2==2.9.1
pycairo==1.20.1

PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
python-dateutil==2.7.3
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.7.2
wadllib==1.3.3
yarl==1.6.3

celery==5.1.2
django-heroku==0.3.1
django-environ==0.8.1
django-redis==5.0.0
drf-yasg==1.20.0
django-cors-headers==3.10.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
flower==1.0.0

Faker==9.7.1
mock==4.0.3